backend/requirements.txt
# =============================================================================
# PHARMA RAG - DEPENDENCIES
# =============================================================================

# -----------------------------------------------------------------------------
# FASTAPI & SERVER
# -----------------------------------------------------------------------------
fastapi
uvicorn
pydantic

# -----------------------------------------------------------------------------
# LANGCHAIN CORE
# -----------------------------------------------------------------------------
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
langchain-experimental

# -----------------------------------------------------------------------------
# VECTOR STORE
# -----------------------------------------------------------------------------
chromadb

# -----------------------------------------------------------------------------
# EMBEDDINGS & RERANKER
# -----------------------------------------------------------------------------
sentence-transformers

# -----------------------------------------------------------------------------
# PDF LOADING
# -----------------------------------------------------------------------------
PyMuPDF

# -----------------------------------------------------------------------------
# BM25
# -----------------------------------------------------------------------------
rank-bm25

# -----------------------------------------------------------------------------
# ENVIRONMENT
# -----------------------------------------------------------------------------
python-dotenv

# -----------------------------------------------------------------------------
# FRONTEND
# -----------------------------------------------------------------------------
streamlit